qc status logic update
