+++ src/main.rs
@@ -369,9 +369,10 @@ async fn main() -> Result<()> {
                                 .ok()
                                 .map(|v| v.into_iter().map(|(c, _)| c).collect::<Vec<_>>());
                             let git_status = git_info.status()?;
-                            let qc_status =
-                                QCStatus::determine_status(&issue_thread, &git_status, &git_info)
-                                    .await?;
+                            let qc_status = QCStatus::determine_status(
+                                &issue_thread,
+                                file_commits.as_ref().unwrap_or(&Vec::new()),
+                            )?;
                             println!(
                                 "{}",
                                 single_issue_status(
